+++ packages/elementary_tools/elementary_cli/README.md
@@ -14,7 +14,7 @@ Run `flutter pub global activate elementary_cli`
 
 ## Integration
 The only option for now is exit codes. You can find full list of exceptions and exit codes
-in [exit_code_exception.dart](lib/exit_code_exception.dart) file.
+in [exit_code_exception.dart](lib/utils/exit_code_exception.dart) file.
 
 ## Resources
 * [dart server docs](https://htmlpreview.github.io/?https://github.com/dart-lang/sdk/blob/main/pkg/analysis_server/doc/api.html)
