+++ packages/elementary_tools/elementary_cli/lib/commands/generate/generate_module.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:elementary_cli/commands/generate/generate.dart';
-import 'package:elementary_cli/exit_code_exception.dart';
-import 'package:elementary_cli/logger.dart';
+import 'package:elementary_cli/utils/exit_code_exception.dart';
+import 'package:elementary_cli/utils/logger.dart';
 import 'package:path/path.dart' as p;
 
 /// `elementary_tools generate module` command
