+++ packages/elementary_tools/elementary_cli/lib/commands/generate/generate_test/generate_test_wm.dart
@@ -4,8 +4,8 @@ import 'dart:io';
 import 'package:args/args.dart';
 import 'package:elementary_cli/analysis_client/elementary_client.dart';
 import 'package:elementary_cli/commands/generate/generate.dart';
-import 'package:elementary_cli/exit_code_exception.dart';
-import 'package:elementary_cli/logger.dart';
+import 'package:elementary_cli/utils/exit_code_exception.dart';
+import 'package:elementary_cli/utils/logger.dart';
 import 'package:path/path.dart' as p;
 
 /// `elementary_tools generate test wm` command
