--- packages/elementary_tools/elementary_cli/lib/console_writer.dart
@@ -1,8 +1,0 @@
-// ignore_for_file: avoid_print
-
-// // Writes messages to console
-// class ConsoleWriter {
-//   static void write(Object? object) {
-//     print(object);
-//   }
-// }
