+++ packages/elementary_tools/elementary_cli/lib/utils/exit_code_exception.dart
