+++ packages/elementary_tools/elementary_cli/lib/utils/logger.dart
